+++ src/comps/Filter/FilterSummaryActive.svelte
@@ -8,29 +8,30 @@
 </script>
 
 {#if !!activeStr || activeIds.length > 0}
-  <b>{label}:</b>
   {#if !!activeStr}
-    <span class="filter__item link" on:click={e => onClick(e, 'search')}>
-      <i class="small link">delete</i>
-      <span>"{activeStr}"</span>
-      <div class="tooltip bottom small-margin">lösche "{activeStr}"</div>
+    <span class="chip round small link white-text no-margin" on:click={e => onClick(e, 'search')}>
+      <i class="small">delete</i>
+      <span>{label}: "{activeStr}"</span>
+      <div class="tooltip bottom small-margin">entferne {label}: "{activeStr}"</div>
     </span>
   {:else if activeIds.length > 0}
     {#each activeIds as itemId}
-      <span class="filter__item link" on:click={e => onClick(e, itemId)}>
-        <i class="small link">delete</i>
+      <span class="chip round small link white-text no-margin" on:click={e => onClick(e, itemId)}>
+        <i class="small">delete</i>
         <span>{store.filter(item => item.id === itemId)[0].de}</span>
-        <div class="tooltip bottom small-margin">lösche Filter</div>
+        <div class="tooltip bottom small-margin">
+          entferne {label}: {store.filter(item => item.id === itemId)[0].de}
+        </div>
       </span>
     {/each}
   {/if}
 {/if}
 
 {#if reset}
-  <span on:click={e => onClick(e, 'all')} class="filter__item link">
-    <i class="small link">delete</i>
+  <span class="chip round small link white-text no-margin" on:click={e => onClick(e, 'all')}>
+    <i class="small">delete</i>
     <span>{label}</span>
-    <div class="tooltip bottom small-margin">Lösche alle Filter</div>
+    <div class="tooltip bottom small-margin">entferne {label}</div>
   </span>
 {/if}