+++ src/stores/states.ts
@@ -1,45 +1,5 @@
 import { writable } from 'svelte/store';
-import { UNLOADED } from '@interfaces';
 import { capitalizeFirstLetter } from '@utils';
-
-const storedActiveSelectionWritable = writable({
-  loadedData: {
-    changes: UNLOADED,
-    inst: UNLOADED,
-    movie: UNLOADED,
-    history: UNLOADED,
-  },
-  lastCursor: {
-    hash: '',
-    dateStr: '',
-  },
-  tags: [],
-  parts: [],
-  partTypes: [],
-  colors: [],
-  states: [],
-  search: '',
-  reason: 'default',
-  page: '',
-});
-export const storedActiveSelection = {
-  subscribe: storedActiveSelectionWritable.subscribe,
-  set: storedActiveSelectionWritable.set,
-  update: storedActiveSelectionWritable.update,
-};
-
-const storedActiveProductWritable = writable({
-  product: {
-    id: 0,
-    type: '', // product-type todaychanges, hearts etc.
-  },
-  reason: 'default',
-});
-export const storedActiveProduct = {
-  subscribe: storedActiveProductWritable.subscribe,
-  set: storedActiveProductWritable.set,
-  update: storedActiveProductWritable.update,
-};
 
 import statesJSON from '../../data/states.json';
 
@@ -52,6 +12,7 @@ export const sortedStates = statesJSON.en.map((state, i) => {
     api: statesJSON.api[i],
   };
 });
+
 const storedStatesWritable = writable(sortedStates);
 export const storedStates = {
   subscribe: storedStatesWritable.subscribe,